# client/src/main/resources/openapi/resos-openapi.yml
openapi: 3.0.0
info:
  title: ResOS API
  version: '1.2'
  description: |
    ResOS API for managing restaurant bookings, tables, opening hours, orders, feedback and customers.
    
    For support or feature requests, contact: hi@resos.com
  contact:
    email: hi@resos.com
servers:
  - url: https://api.resos.com/v1
    description: Production API Server
security:
  - basicAuth: []
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: |
        Authentication using API key as username (no password required).
        The API key must be base64 encoded with a trailing colon.
  schemas:
...
    Customer:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
          pattern: '^\+[0-9]+$'
        createdAt:
          type: string
          format: date-time
        lastBookingAt:
          type: string
          format: date-time
        bookingCount:
          type: integer
        totalSpent:
          type: number
          format: float
        metadata:
          type: object
          description: "Custom metadata associated with the customer"
    OpeningHours:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        date:
          type: string
          format: date
        opens:
          type: string
          pattern: '^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$'
        closes:
          type: string
          pattern: '^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$'
        isOpen:
          type: boolean
        maxPeople:
          type: integer
        metadata:
          type: object
...
    Feedback:
      type: object
      properties:
        _id:
          type: string
        bookingId:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
        customer:
          $ref: '#/components/schemas/Customer'
        isPublic:
          type: boolean
          default: false
...
  parameters:
...
    limit:
      name: limit
      in: query
      description: Number of records to return (max 100)
      required: false
      schema:
        type: integer
        default: 100
        maximum: 100
    skip:
      name: skip
      in: query
      description: Number of records to skip
      required: false
      schema:
        type: integer
        default: 0
    sort:
      name: sort
      in: query
      description: Sort field and direction (1 ascending, -1 descending)
      required: false
      schema:
        type: string
        example: "createdAt:1"
    customQuery:
      name: customQuery
      in: query
      description: Search expression for filtering results
      required: false
      schema:
        type: string
        example: "people>1,duration>60"
paths:
...
  /customers:
    get:
      summary: List customers
      description: Retrieve a list of customers
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/skip'
        - $ref: '#/components/parameters/sort'
        - $ref: '#/components/parameters/customQuery'
      responses:
        '200':
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
  /customers/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Get customer
      description: Retrieve a specific customer by ID
      responses:
        '200':
          description: Customer details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          description: Customer not found
...
  /opening-hours/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Get opening hours
      description: Retrieve specific opening hours by ID
      responses:
        '200':
          description: Opening hours details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpeningHours'
        '404':
          description: Opening hours not found
...
  /feedback:
    get:
      summary: List feedback
      description: Retrieve a list of feedback
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/skip'
        - $ref: '#/components/parameters/sort'
        - $ref: '#/components/parameters/customQuery'
      responses:
        '200':
          description: List of feedback
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Feedback'
  /feedback/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Get feedback
      description: Retrieve specific feedback by ID
      responses:
        '200':
          description: Feedback details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Feedback'
        '404':
          description: Feedback not found
...